pyproject.toml
[tool.poetry]
name = "bookr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
drf-spectacular = "^0.26.4"
pendulum = "^2.1.2"
drf-nested-routers = "^0.93.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ipython = "^8.14.0"


[tool.poetry.group.local.dependencies]
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

